requirements.txt
python-dotenv
python-telegram-bot
tabulate